+++ CHANGELOG.md
@@ -31,7 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - A new global uniform IsGPUParticlesEnabled was added into constants CB/UBO which equals to 1 when GPU particles simulation is enabled in the project.
 
-- GPU particle systems now support forces.
+- GPU particle systems now support forces and all particles are sorted on GPU using bitonic algorithm.
 
 ### Changed